requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-extensions==3.2.1
django-gsheets==0.0.10
django-rest-passwordreset==1.3.0
django4-background-tasks==1.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.6
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
